Web crawler/requirements.txt
beautifulsoup4~=4.9.3
requests~=2.24.0
mysql-connector-python~=8.0.18
numpy~=1.19.1
matplotlib~=3.3.1
pillow~=7.2.0